Created Static Class PathGeeseFly & Passed All Tests!

+++ LoggingKata.Test/TacoParserTests.cs
@@ -16,13 +16,12 @@ namespace LoggingKata.Test
 
             //Assert
             Assert.NotNull(actual);
-
         }
 
         [Theory]
         [InlineData("34.073638, -84.677017, Taco Bell Acwort...", -84.677017)]
         //Add additional inline data. Refer to your CSV file.
-        public void ShouldParseLongitude(string line, decimal expected)
+        public void ShouldParseLongitude(string line, double expected)
         {
             // TODO: Complete the test with Arrange, Act, Assert steps below.
             //       Note: "line" string represents input data we will Parse 
@@ -44,7 +43,7 @@ namespace LoggingKata.Test
         //TODO: Create a test called ShouldParseLatitude
         [Theory]
         [InlineData("34.073638, -84.677017, Taco Bell Acwort...", 34.073638)]
-        public void ShouldParseLatitude(string line, decimal expected)
+        public void ShouldParseLatitude(string line, double expected)
         {
             //Arrange
             TacoParser tacoParser = new TacoParser();
